Implemented remaining constructors for all classes based on UML Diagram

+++ src/main/java/com/CMSC447/nurseroster/domain/HoursRequired.java
@@ -11,10 +11,10 @@ import java.util.List;
 public class HoursRequired implements PersonalConstraint{
    
 	int id; 
-	int amount; 
+	double amount; 
 	String per; 
 	
-	HoursRequired(int id, int amount, String per){
+	HoursRequired(int id, double amount, String per){
 		this.id = id;
 		this.amount = amount;
 		this.per = per;
